Bf16 (#190)
* bf16 on Ampere

* python dtype

+++ python/mirage/kernel.py
@@ -237,7 +237,7 @@ class KNGraph:
 
         output_shapes = self._cached_results["output_shapes"]
         output_tensors = [
-            torch.zeros(shape, dtype=torch.float16, device=input_tensors[0].device) for shape in output_shapes
+            torch.zeros(shape, dtype=input_tensors[0].dtype, device=input_tensors[0].device) for shape in output_shapes
         ]
         if(verbose):
             print("Input tensors:")
@@ -272,7 +272,7 @@ class KNGraph:
                 meta["shape"],
                 meta["strides"],
                 device=input_tensors[0].device,
-                dtype=torch.float16,
+                dtype=input_tensors[0].dtype,
             )
             for meta in results["output_directives"]
         ]
@@ -444,7 +444,7 @@ class KNGraph:
                             for t in dtensors:
                                 dims = [t.dim(i) for i in range(t.num_dims)]
                                 input_tensors.append(
-                                    torch.randn(dims, dtype=torch.float16, device="cuda:0")
+                                    torch.randn(dims, dtype=t.dtype, device="cuda:0")
                                 )
                             starter = torch.cuda.Event(enable_timing=True)
                             ender = torch.cuda.Event(enable_timing=True)
@@ -458,7 +458,7 @@ class KNGraph:
                     for t in dtensors:
                         dims = [t.dim(i) for i in range(t.num_dims)]
                         input_tensors.append(
-                            torch.randn(dims, dtype=torch.float16, device="cuda:0")
+                            torch.randn(dims, dtype=t.dtype, device="cuda:0")
                         )
                     starter = torch.cuda.Event(enable_timing=True)
                     ender = torch.cuda.Event(enable_timing=True)
@@ -472,7 +472,7 @@ class KNGraph:
                 for t in dtensors:
                     dims = [t.dim(i) for i in range(t.num_dims)]
                     input_tensors.append(
-                        torch.randn(dims, dtype=torch.float16, device="cuda:0")
+                        torch.randn(dims, dtype=t.dtype, device="cuda:0")
                     )
                 starter = torch.cuda.Event(enable_timing=True)
                 ender = torch.cuda.Event(enable_timing=True)
